requirements.txt
google-generativeai
agentkit
flask